+++ tests/pure/web.structured-clone.js
@@ -21,6 +21,8 @@ QUnit.module('structuredClone', () => {
     assert.name(structuredClone, 'structuredClone');
     assert.arity(structuredClone, 1);
     assert.throws(() => structuredClone(), 'throws without arguments');
+    assert.same(structuredClone(1, null), 1, 'null as options');
+    assert.same(structuredClone(1, undefined), 1, 'undefined as options');
   });
 
   function cloneTest(value, verifyFunc) {